third_party/blink/web_tests/fast/innerHTML/innerHTML-svg-write-expected.txt

Test that setting innerHTML and outerHTML properties works on SVG elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS svg.innerHTML is "<rect></rect>"
PASS svg.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS svg.innerHTML is "<foreignObject><svg></svg></foreignObject>"
PASS svg.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS svg.firstChild.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS svg.innerHTML is "<foreignObject><p>x</p></foreignObject>"
PASS svg.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS svg.firstChild.firstChild.namespaceURI is "http://www.w3.org/1999/xhtml"
PASS svg.innerHTML is "<circle></circle>"
PASS svg.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS successfullyParsed is true

TEST COMPLETE
